refactor(device): split device path protocol tools into separate modules.
