+++ src/runtime/protocol/mod.rs
@@ -1,5 +1,5 @@
 pub(crate) mod block_io;
-pub(crate) mod device_path;
+pub(crate) mod device;
 pub(crate) mod fs;
 pub(crate) mod graphics_output;
 pub(crate) mod handle;
